fix #643
